Allow definition of CP_USE_DOUBLES to force use of doubles. Add flag to .cabal file.
Want to ensure that double precisions are always used so that
the Haskell library runs the same on all platforms.
chipmunk_types.h contained a bug where if TARGET_OS_IPHONE
was defined but __LP64__ was not defined, then CP_USE_DOUBLES
would be redefined no matter whether it was defined or not.
If it is defined now then it will not be redefined.