Also defined __LP64__. I don't know why this works or is necessary.
This project is being used in a game of mine called Epidemic.
Not defining __LP64__ did not produce the right behaviour on the iOS Simulator!
The germs moved around in very strange ways and the game quickly crashed.
So, now I'm defining `__LP64__` even though I don't know why it works!

It seems to me like it shouldn't since I'm building for the i386 architecture.
How can saying that longs and pointers are 64 bits (which is what `__LP64__` does) work?